# procfile: more comfortable initial settings and faster/fewer reallocs